+++ src/vak/prep/__init__.py
@@ -1,6 +1,7 @@
 from . import (
     audio_dataset,
     constants,
+    dataset_df,
     frame_classification,
     spectrogram_dataset,
 )
@@ -10,6 +11,7 @@ from .prep import prep
 __all__ = [
     'audio_dataset',
     'constants',
+    'dataset_df',
     'frame_classification',
     'prep',
     'spectrogram_dataset',
