+++ src/vak/prep/dataset_df_helper.py
@@ -1,0 +1,119 @@
+"""Helper functions for working with datasets represented as a pandas.DataFrame"""
+import numpy as np
+import pandas as pd
+
+
+def get_dataset_csv_filename(data_dir_name: str, timenow: str) -> str:
+    """Get name of csv file representing dataset.
+
+    This function is called by
+    :func:`vak.prep.frame_classification.dataset_df.get_dataset_csv_path`.
+
+    Parameters
+    ----------
+    data_dir_name : str
+        Name of directory specified as parameter ``data_dir``
+        when calling :func:`vak.core.prep.prep`.
+        This becomes the "prefix" of the csv filename.
+    timenow : str
+        Timestamp.
+        This becomes the "suffix" of the csv filename.
+
+    Returns
+    -------
+    dataset_csv_filename : str
+        String, in the form f"{data_dir_name}_prep_{timenow}.csv"
+    """
+    return f"{data_dir_name}_prep_{timenow}.csv"
+
+
+def get_dataset_csv_path(dataset_path: pathlib.Path, data_dir_name: str, timenow: str) -> pathlib.Path:
+    """Returns the path that should be used to save
+    a pandas DataFrame representing a dataset
+    to a csv file.
+
+    Parameters
+    ----------
+    dataset_path : str, pathlib.Path
+        Path to directory that represents dataset.
+    data_dir_name : str
+        Name of directory specified as parameter ``data_dir``
+        when calling :func:`vak.core.prep.prep`.
+        This becomes the "prefix" of the csv filename.
+    timenow : str
+        Timestamp.
+        This becomes the "suffix" of the csv filename.
+
+    Returns
+    -------
+    dataset_csv_path : pathlib.Path
+        Path that is used when saving ``dataset_df`` as a csv file
+        in the root of the dataset directory, ``dataset_path``.
+    """
+    dataset_csv_filename = get_dataset_csv_filename(data_dir_name, timenow)
+    dataset_csv_path = dataset_path / dataset_csv_filename
+    return dataset_csv_path
+
+
+def add_split_col(df: pd.DataFrame, split: str) -> pd.DataFrame:
+    """Add a 'split' column to a pandas DataFrame.
+
+    Used by :func:`vak.prep`
+    to assign an entire dataset to the same split,
+    e.g. 'train' or 'predict'.
+    All rows in the 'split' column will have the value specified.
+
+    Parameters
+    ----------
+    df : pandas.DataFrame
+        A dataframe that represents a dataset.
+    split : str
+        A string that will be assigned to every row
+        in the added "split" column.
+        One of {'train', 'val', 'test', 'predict'}.
+    """
+    if split not in {"train", "val", "test", "predict"}:
+        raise ValueError(
+            f"value for split should be one of {{'train', 'val', 'test', 'predict'}}, but was '{split}'"
+        )
+    split_col = np.asarray([split for _ in range(len(df))], dtype="object")
+    df["split"] = split_col
+    return df
+
+
+def validate_and_get_timebin_dur(df: pd.DataFrame, expected_timebin_dur: float | None = None) -> float:
+    """Validate that there is a single, unique value for the time bin duration of all
+    spectrograms in a dataset. If so, return that value.
+
+    The dataset is represented as a pandas DataFrame.
+
+    Parameters
+    ----------
+    df : pandas.Dataframe
+        A pandas.DataFrame created by
+        ``vak.io.dataframe.from_files``
+        or ``vak.io.spect.to_dataframe``.
+    expected_timebin_dur : float
+
+    Returns
+    -------
+    timebin_dur : float
+        The duration of a time bin in seconds
+        for all spectrograms in the dataset.
+    """
+    timebin_dur = df["timebin_dur"].unique()
+    if len(timebin_dur) > 1:
+        raise ValueError(
+            f"Found more than one time bin duration in dataset: {timebin_dur}"
+        )
+    elif len(timebin_dur) == 1:
+        timebin_dur = timebin_dur.item()
+
+    if expected_timebin_dur:
+        if timebin_dur != expected_timebin_dur:
+            raise ValueError(
+                f"Timebin duration from dataset, {timebin_dur}, "
+                f"did not match expected timebin duration, {expected_timebin_dur}."
+            )
+
+    return timebin_dur
